+++ spec/dummy_rails_app/app/channels/tailwind_channel.rb
@@ -1,0 +1,10 @@
+# Test channel to ensure TailwindCSS compiler channel works
+class TailwindChannel < ApplicationCable::Channel
+  def subscribed
+    stream_from "tailwind_updates"
+  end
+  
+  def unsubscribed
+    # Any cleanup needed when channel is unsubscribed
+  end
+end