+++ spec/dummy_rails_app/config/routes.rb
@@ -1,0 +1,6 @@
+Rails.application.routes.draw do
+  mount ActionCable.server => "/cable"
+  
+  resources :pages, only: [:index, :show]
+  root "pages#index"
+end